+++ MANIFEST.in
@@ -1,4 +1,4 @@
 include LICENSE
 include README.md
 
-recursive-include genpy_stubgen *.py
+recursive-include genpyi *.py
