+++ README.md
@@ -1,11 +1,11 @@
-# genpy\_stubgen
+# genpyi
 
 A Python stub generator from genmsg specs
 
 ## Installation
 
 ```sh
-pip install genpy-stubgen
+pip install genpyi
 ```
 
 ## Usage
@@ -32,8 +32,8 @@ Examples:
 ### CLI
 
 ```
-$ genpy_stubgen --help
-Usage: genpy_stubgen [-h] {msg,srv,module} ...
+$ genpyi --help
+Usage: genpyi [-h] {msg,srv,module} ...
 positional arguments:
   {msg,srv,module}
     msg             Generate stub files from .msg files
@@ -47,24 +47,24 @@ optional arguments:
 Examples:
 ```sh
 # Message files
-$ genpy_stubgen msg custom_msgs custom_msgs/msg/Custom.msg
-$ genpy_stubgen msg std_msgs --out-dir out /opt/ros/melodic/share/std_msgs/msg/Header.msg
-$ genpy_stubgen msg sensor_msgs --out-dir out \
+$ genpyi msg custom_msgs custom_msgs/msg/Custom.msg
+$ genpyi msg std_msgs --out-dir out /opt/ros/melodic/share/std_msgs/msg/Header.msg
+$ genpyi msg sensor_msgs --out-dir out \
     -Istd_msgs:/opt/ros/melodic/share/std_msgs/msg \
     -Isensor_msgs:/opt/ros/melodic/share/sensor_msgs/msg \
     /opt/ros/melodic/share/sensor_msgs/msg/PointCloud2.msg
 
 # Service files
-$ genpy_stubgen srv custom_msgs custom_msgs/srv/Custom.msg
-$ genpy_stubgen srv nav_msgs --out-dir out \
+$ genpyi srv custom_msgs custom_msgs/srv/Custom.msg
+$ genpyi srv nav_msgs --out-dir out \
     -Istd_msgs:/opt/ros/melodic/share/std_msgs/msg \
     -Isensor_msgs:/opt/ros/melodic/share/sensor_msgs/msg \
     /opt/ros/melodic/share/sensor_msgs/srv/SetCameraInfo.srv
 ```
 
-- `genpy_stubgen msg` / `genpy_stubgen srv`
+- `genpyi msg` / `genpyi srv`
   ```sh
-  Usage: genpy_stubgen {msg,srv} [-h] [--out-dir OUT_DIR]
+  Usage: genpyi {msg,srv} [-h] [--out-dir OUT_DIR]
                                  [--include-path INCLUDE_PATH]
                                  package files [files ...]
 
@@ -80,9 +80,9 @@ $ genpy_stubgen srv nav_msgs --out-dir out \
     --include-path INCLUDE_PATH, -I INCLUDE_PATH
                           Include paths for processing given files
   ```
-- `genpy_stubgen module`
+- `genpyi module`
   ```sh
-  Usage: genpy_stubgen module [-h] [--out-dir OUT_DIR] package_dir
+  Usage: genpyi module [-h] [--out-dir OUT_DIR] package_dir
 
   Positional arguments:
     package_dir        Package directory to create __init__.pyi
