+++ genpyi/_compat.py
@@ -1,7 +1,7 @@
 import functools
 import sys
 
-from genpy_stubgen._typing import TYPE_CHECKING
+from ._typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from typing import Any, Callable, Dict, TypeVar
