+++ pyproject.toml
@@ -8,7 +8,7 @@ enable_isort = true
 enable_mypy = true
 mypy_preset = "strict"
 py_version = "py27"
-isort_known_first_party = ["genpy_stubgen"]
+isort_known_first_party = ["genpyi"]
 [tool.pysen.lint.source]
   excludes = ["tests/integration_tests/expected"]
 [[tool.pysen.lint.mypy_targets]]
