+++ scripts/run_genpyi.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from genpy_stubgen.cli import cli
+from genpyi.cli import cli
 
 
 if __name__ == "__main__":
