+++ tests/integration_tests/test_modules.py
@@ -2,7 +2,7 @@ import os
 import tempfile
 from typing import Dict, List
 
-from genpy_stubgen.cli import run_module_stubgen
+from genpyi.cli import run_module_stubgen
 
 from .utils import assert_output_equals, message_path, service_path
